requirements.txt
apache-tvm==0.11.1
astor==0.8.1
astunparse==1.6.3
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
cmake==3.26.3
coloredlogs==15.0.1
dataclasses==0.6
decorator==5.1.1
filelock==3.12.0
flatbuffers==23.5.9
graphviz==0.20.1
h5py==3.8.0
humanfriendly==10.0
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
lit==16.0.5
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnx==1.14.0
onnxoptimizer==0.3.13
onnxruntime==1.14.1
packaging==23.1
Pillow==9.5.0
protobuf==4.23.1
psutil==5.9.5
pycg==0.0.6
python-scalpel==1.0b0
regex==2023.5.5
requests==2.30.0
scipy==1.10.1
six==1.16.0
sympy==1.12
synr==0.6.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
triton==2.0.0
typed-ast==1.5.4
typeshed-client==2.3.0
typing_extensions==4.5.0
urllib3==2.0.2
zipp==3.15.0
